Rename postRating
